1.2.1 add refresh patch for mech lab chassis storage view on triggered builds
