+++ SalvageOperations/mod.json
@@ -2,7 +2,7 @@
     "Name": "SalvageOperations",
     "DLL": "SalvageOperations.dll",
 
-    "Version": "1.2.0",
+    "Version": "1.2.1",
     "Description": "Changes how mech salvage works.",
     "Author": "mpstark",
     "Website": "www.github.com/BattleTechModders/SalvageOperations",
